feat: book records to access history
